Implemented a way to hide/show the notes section
This is important for user who want a handled way of showing/hiding such
section. Moreover, it also allows you to hide this section from the very
beginning, so this mode can also be used by people who only want to text itself
and maybe the outline (i.e. remove what it could be perceived as clutter).

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>