writer-org: add the idea for a main file
When having multiple org files which include one another, in most cases there is
a file that groups everything, acting as a main file. Now you can specify the
main file, so wherever you are editing, you can call `writer-org-export-to-pdf'
and the main file will be the one picked when exporting instead of the current one.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>